feat: Expose a new Bar integrated into React Tree
BREAKING CHANGE : You need to use `BarComponent` into your application if you want to have the cozy-bar
